+++ scripts/events/abilitySave.js
@@ -266,6 +266,7 @@ async function rollSave(wrapped, config, dialog = {}, message = {}) {
         messageData.template = 'modules/midi-qol/templates/roll-base.html';
         await returnData.toMessage(messageData, {rollMode: returnData.options?.rollMode ?? rollMode});
     }
+    await executeMacroPass(this, 'post', saveId, options, returnData, config, dialog, message);
     return shouldBeArray ? [returnData] : returnData;
 }
 function patch() {